+++ test/com/brein/activity/TestActivity.java
@@ -263,14 +263,10 @@ public class TestActivity {
     public void testResetAllValues() {
 
         final BreinActivity breinActivity = new BreinActivity();
-        breinActivity.setAdditionalUrl("www.test.com.au");
         breinActivity.setBreinActivityType(BreinActivityType.ADD_TO_CART);
         breinActivity.setBreinCategoryType(BreinCategoryType.EDUCATION);
         breinActivity.setDescription("Description");
-        breinActivity.setIpAddress("10.11.222.333");
-        breinActivity.setReferrer("Referrer");
-        breinActivity.setSessionId("777ABBG");
-        breinActivity.setUserAgent("userAgentValue");
+
         final Map<String, Object> tagMap = new HashMap<>();
         tagMap.put("t1", "0.0");
         breinActivity.setTagsMap(tagMap);
@@ -281,14 +277,9 @@ public class TestActivity {
 
         // first check init
         breinActivity.init();
-        assertTrue("addtionalUrl is not empty", breinActivity.getAdditionalUrl().equals(""));
         assertTrue("breinActivityType is not empty", breinActivity.getBreinActivityType().equals(""));
         assertTrue("breinCategoryType is not empty", breinActivity.getBreinCategoryType().equals(""));
         assertTrue("description is not empty", breinActivity.getDescription().equals(""));
-        assertTrue("ipAddress is not empty", breinActivity.getIpAddress().equals(""));
-        assertTrue("referrer is not empty", breinActivity.getReferrer().equals(""));
-        assertTrue("sessionId is not empty", breinActivity.getSessionId().equals(""));
-        assertTrue("userAgent is not empty", breinActivity.getUserAgent().equals(""));
         assertTrue("tags lib is not null", breinActivity.getTagsMap() == null);
 
         // breinUser and breinConfig should be valid
@@ -302,5 +293,4 @@ public class TestActivity {
         assertTrue("breinUser is not null!", breinActivity.getBreinUser() == null);
         assertTrue("breinConfig is not null!", breinActivity.getConfig() == null);
     }
-
 }